xbel parsing in place
Here is where I am:

* I download the existing xbel file, parse it and create the bookmarks
in the map.

* I have not dealt with the tabs/spaces issue.

* please consider something like:
https://www.npmjs.com/package/js-beautify
to keep us to a coding style.

* This works except on the 2nd sync I get the following error:
Syncing failed with Trying to create a URL that is already bookmarked.
I do not know what I did wrong.

*This is still based on 2.2.3, after the last rebase causing merge
conflicts I wanted to leave that alonw.

*The parsing (and xml generation) code could just as easily work with
the bookmarks.html file that chrome and probably firefox already work
with. So we can either offer a checkbox for the appropriate format or
change to the html mechanism if we wanted to.

*further, when this gets more solid, we need to think of assigning
folders ids as well, they mean something in xbel and in html formats.
It will also help with bookmark ordering.